requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
matplotlib>=3.7.0
pillow>=9.5.0
pyyaml>=6.0.0
tqdm>=4.65.0
pyldpc>=0.1.8
